Add files via upload
Parse_function attempts to parse a function and return a ParseResult with a Function_Expression as the value and the position of the next token as nextpos. If it is not immediately obvious that what is being parsed is a function then it calls Parse_Expression and attempts to parse an expression. If it seems like what is being parsed is a function then we attempt to parse the type, identifier, parameters, body, and return statement. An error is erected when the token read is not the token expected.